+++ CHANGELOG.md
@@ -6,8 +6,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2023-08-16
 
-## [0.1.0] - 26/10/2022
+### Added
+
+- Add a command to upload binaries directly to object storage.
+- Add a command to delete samples from both DB and object storage.
+- Add a command to list users and files from both DB and object storage.
+- Add option to download binaries directly from object storage.
+- Add a command to generate malware souk families.
+- Chore: Github build action workflow.
+
+
+## [0.1.0] - 26-10-2022
 
 ### Added